# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MoveProfileJobStatus = typeof MoveProfileJobStatus[keyof typeof MoveProfileJobStatus];


export const MoveProfileJobStatus = {
  offer: 'offer',
  'no-offer': 'no-offer',
} as const;

export type MoveProfileFamilyStatus = typeof MoveProfileFamilyStatus[keyof typeof MoveProfileFamilyStatus];


export const MoveProfileFamilyStatus = {
  alone: 'alone',
  others: 'others',
} as const;

export type MoveProfileArrivalStatus = typeof MoveProfileArrivalStatus[keyof typeof MoveProfileArrivalStatus];


export const MoveProfileArrivalStatus = {
  date: 'date',
  visa: 'visa',
} as const;

export interface MoveProfile {
  name: string;
  nationality: string;
  jobStatus: MoveProfileJobStatus;
  salaryRange: string;
  familyStatus: MoveProfileFamilyStatus;
  /** @minimum 1 */
  adults: number;
  /** @minimum 0 */
  children: number;
  arrivalStatus: MoveProfileArrivalStatus;
  /** @nullable */
  arrivalDate: string | null;
  /** @nullable */
  currentCity: string | null;
}

export interface ExtractProfileInput {
  /** @minLength 1 */
  text: string;
  answers?: MoveProfile;
}

export interface GenerateMovePlanInput {
  profile: MoveProfile;
}

export type MoveTaskPriority = typeof MoveTaskPriority[keyof typeof MoveTaskPriority];


export const MoveTaskPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type MoveTaskCategory = typeof MoveTaskCategory[keyof typeof MoveTaskCategory];


export const MoveTaskCategory = {
  visa: 'visa',
  housing: 'housing',
  job: 'job',
  bank: 'bank',
  'id-medical': 'id-medical',
  community: 'community',
  sim: 'sim',
} as const;

export interface MoveTask {
  id: string;
  title: string;
  description: string;
  priority: MoveTaskPriority;
  category: MoveTaskCategory;
  /**
     * @minimum 0
     * @maximum 100
     */
  scorePoints: number;
  officialSource: boolean;
}

export interface CostLine {
  category: string;
  /** @minimum 0 */
  lowAed: number;
  /** @minimum 0 */
  highAed: number;
}

export interface MovePlan {
  beforeArrival: MoveTask[];
  week1: MoveTask[];
  week2: MoveTask[];
  /** @minimum 0 */
  estimatedFirstMonthCost: number;
  costBreakdown: CostLine[];
  costNote: string;
}

export type MoveMateChatInputConversationHistoryItemRole = typeof MoveMateChatInputConversationHistoryItemRole[keyof typeof MoveMateChatInputConversationHistoryItemRole];


export const MoveMateChatInputConversationHistoryItemRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export type MoveMateChatInputConversationHistoryItem = {
  role: MoveMateChatInputConversationHistoryItemRole;
  text: string;
};

export interface MoveMateChatInput {
  /** @minLength 1 */
  message: string;
  profile: MoveProfile;
  plan: MovePlan;
  completedTaskIds: string[];
  /** @maxItems 12 */
  conversationHistory?: MoveMateChatInputConversationHistoryItem[];
}

export type MoveMateChatResponseIntent = typeof MoveMateChatResponseIntent[keyof typeof MoveMateChatResponseIntent];


export const MoveMateChatResponseIntent = {
  'next-step': 'next-step',
  bank: 'bank',
  visa: 'visa',
  'task-completed': 'task-completed',
  general: 'general',
} as const;

export interface MoveMateChatResponse {
  reply: string;
  aiPowered: boolean;
  intent: MoveMateChatResponseIntent;
  completedTaskIds: string[];
  nextTask: MoveTask | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  settlementScore: number;
}