# lib/api-zod/src/generated/types/costLine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CostLine {
  category: string;
  /** @minimum 0 */
  lowAed: number;
  /** @minimum 0 */
  highAed: number;
}