# lib/api-zod/src/generated/types/generateMovePlanInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MoveProfile } from './moveProfile';

export interface GenerateMovePlanInput {
  profile: MoveProfile;
}