# lib/api-zod/src/generated/types/moveMateChatInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MoveMateChatInputConversationHistoryItem } from './moveMateChatInputConversationHistoryItem';
import type { MovePlan } from './movePlan';
import type { MoveProfile } from './moveProfile';

export interface MoveMateChatInput {
  /** @minLength 1 */
  message: string;
  profile: MoveProfile;
  plan: MovePlan;
  completedTaskIds: string[];
  /** @maxItems 12 */
  conversationHistory?: MoveMateChatInputConversationHistoryItem[];
}