# lib/api-zod/src/generated/types/moveMateChatResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MoveMateChatResponseIntent } from './moveMateChatResponseIntent';
import type { MoveTask } from './moveTask';

export interface MoveMateChatResponse {
  reply: string;
  aiPowered: boolean;
  intent: MoveMateChatResponseIntent;
  completedTaskIds: string[];
  nextTask: MoveTask | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  settlementScore: number;
}