# lib/api-zod/src/generated/types/movePlan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CostLine } from './costLine';
import type { MoveTask } from './moveTask';

export interface MovePlan {
  beforeArrival: MoveTask[];
  week1: MoveTask[];
  week2: MoveTask[];
  /** @minimum 0 */
  estimatedFirstMonthCost: number;
  costBreakdown: CostLine[];
  costNote: string;
}