# lib/api-zod/src/generated/types/moveProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MoveProfileArrivalStatus } from './moveProfileArrivalStatus';
import type { MoveProfileFamilyStatus } from './moveProfileFamilyStatus';
import type { MoveProfileJobStatus } from './moveProfileJobStatus';

export interface MoveProfile {
  name: string;
  nationality: string;
  jobStatus: MoveProfileJobStatus;
  salaryRange: string;
  familyStatus: MoveProfileFamilyStatus;
  /** @minimum 1 */
  adults: number;
  /** @minimum 0 */
  children: number;
  arrivalStatus: MoveProfileArrivalStatus;
  /** @nullable */
  arrivalDate: string | null;
  /** @nullable */
  currentCity: string | null;
}