# lib/api-zod/src/generated/types/moveTask.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MoveTaskCategory } from './moveTaskCategory';
import type { MoveTaskPriority } from './moveTaskPriority';

export interface MoveTask {
  id: string;
  title: string;
  description: string;
  priority: MoveTaskPriority;
  category: MoveTaskCategory;
  /**
     * @minimum 0
     * @maximum 100
     */
  scorePoints: number;
  officialSource: boolean;
}